requirements.txt
# Core dependencies
boto3==1.34.29
shortuuid==1.0.13
google-genai==1.5.0
httpx>=0.28.1,<1.0.0
google-auth==2.28.0
google-api-core==2.15.0
protobuf==3.20.3

# HTTP and async dependencies
aiohttp==3.9.1
anyio>=4.8.0,<5.0.0
aiosignal>=1.1.2
async-timeout==4.0.3
attrs==23.2.0
exceptiongroup==1.2.0
frozenlist==1.4.1
multidict==6.0.4
yarl==1.9.4
typing-extensions>=4.11.0,<5.0.0dev
sniffio==1.3.0
h11==0.14.0

# Authentication and utilities
certifi==2024.2.2
chardet==5.2.0
idna==3.6
oauthlib==3.2.2
cachetools==5.3.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
google-auth-oauthlib==1.1.0
